+++ .devcontainer/post-install.sh
@@ -1,0 +1,3 @@
+poetry install
+pre-commit install
+pre-commit