+++ strawberry/__init__.py
@@ -9,6 +9,7 @@ from .field import field  # noqa
 from .lazy_type import LazyType  # noqa
 from .mutation import mutation, subscription  # noqa
 from .permission import BasePermission  # noqa
+from .private import Private  # noqa
 from .scalars import ID  # noqa
 from .schema import Schema  # noqa
 from .type import input, interface, type  # noqa