+++ strawberry/http.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, List, Optional
 
 from graphql.error import format_error as format_graphql_error
-from strawberry.schema import ExecutionResult
+from strawberry.types import ExecutionResult
 from typing_extensions import TypedDict