+++ tests/schema/extensions/test_extensions.py
@@ -1,0 +1,98 @@
+import pytest
+
+import strawberry
+from strawberry.extensions import Extension
+
+
+class MyExtension(Extension):
+    def get_results(self):
+        return {"example": "example"}
+
+
+def test_base_extension():
+    @strawberry.type
+    class Person:
+        name: str = "Jess"
+
+    @strawberry.type
+    class Query:
+        @strawberry.field
+        def person(self, info) -> Person:
+            return Person()
+
+    schema = strawberry.Schema(query=Query, extensions=[Extension])
+
+    query = """
+        query {
+            person {
+                name
+            }
+        }
+    """
+
+    result = schema.execute_sync(query)
+
+    assert not result.errors
+
+    assert result.extensions == {}
+
+
+def test_extension():
+    @strawberry.type
+    class Person:
+        name: str = "Jess"
+
+    @strawberry.type
+    class Query:
+        @strawberry.field
+        def person(self, info) -> Person:
+            return Person()
+
+    schema = strawberry.Schema(query=Query, extensions=[MyExtension])
+
+    query = """
+        query {
+            person {
+                name
+            }
+        }
+    """
+
+    result = schema.execute_sync(query)
+
+    assert not result.errors
+
+    assert result.extensions == {
+        "example": "example",
+    }
+
+
+@pytest.mark.asyncio
+async def test_extension_async():
+    @strawberry.type
+    class Person:
+        name: str = "Jess"
+
+    @strawberry.type
+    class Query:
+        @strawberry.field
+        async def person(self, info) -> Person:
+            return Person()
+
+    schema = strawberry.Schema(query=Query, extensions=[MyExtension])
+
+    query = """
+        query {
+            person {
+                name
+            }
+        }
+    """
+
+    result = await schema.execute(query)
+
+    assert not result.errors
+
+    assert result.extensions == {
+        "example": "example",
+    }